abastract EcomProductInfo component
